refactored util func into a class
